requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
PyYAML==6.0.1
nbformat==5.9.2 